FB logout actually fixed
FB's getLoginStatus is needed to get a valid accesstoken to make api calls, but cannot be used to check if a user is currently logged into the app. It only checks if user is logged into facebook and has authenticated your app.

To check log in status check the firebase user profile for the provider.